% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellTypes}
\alias{cellTypes}
\title{GSE827795 data cell type labels}
\format{
An object of class \code{character} of length 367.
}
\usage{
data(gse87795_subset)
}
\description{
A character vector of cell type information for cells in
\code{dat} matrix object
}
\keyword{datasets}
