% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{GSE827795 data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 367 columns.
}
\usage{
data(gse87795_subset)
}
\description{
A subset of expression matrix of GSE827795
}
\keyword{datasets}
