% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matPCs.R
\name{matPCs}
\alias{matPCs}
\title{matPCs function}
\usage{
matPCs(data, percentVar=0.8)
}
\arguments{
\item{data}{Prediction model generated from "multiAdaSampling".}

\item{percentVar}{A dimension reduced matrix.}
}
\value{
Dimensionally reduced matrix.
}
\description{
Performs PCA on a given matrix
}
\details{
This function performs PCA to reduce the dimension of the gene
expression matrix limited from 10 to 20 PCs.
}
\examples{
data("gse87795_subset")

mat.expr = dat

mat.pc = matPCs(mat.expr)

# to capture at least 70\% of overall variability in the dataset,
mat.dim.reduct.70 = matPCs(mat.expr, 0.7)

}
\author{
Pengyi Yang, Taiyun Kim
}
