% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAdaSampling.R
\name{multiAdaSampling}
\alias{multiAdaSampling}
\title{multi Adaptive Sampling function}
\usage{
multiAdaSampling(data, label, classifier="svm", percent=1,
L=10, prob=FALSE, balance=FALSE, iter=3)
}
\arguments{
\item{data}{A dimension reduced matrix.}

\item{label}{A vector of label information for each sample.}

\item{classifier}{Base classifier model, either  "SVM" or "RF"}

\item{percent}{Percentage of samples to select at each iteration during.}

\item{L}{Number of ensembles. Default to 10.}

\item{prob}{logical flag to return sample's probabilities to each class.}

\item{balance}{logical flag to down sample large cell types classes to the
median of all class sizes.}

\item{iter}{A number of iterations to perform adaSampling.}
}
\value{
A final prediction, probabilities for each cell type and the model
are returned as a list.
}
\description{
Performs multiple Adaptive sampling to train a classifier model
}
\examples{
data("gse87795_subset")

mat.expr = dat

mat.pc = matPCs(mat.expr)

result = multiAdaSampling(mat.pc, cellTypes, classifier = "svm",
percent = 1, L = 10)
}
\author{
Pengyi Yang
}
