% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{scReClassify}
\alias{scReClassify}
\title{scReClassify: a package for  post hoc cell type classification of single-cell
RNA-sequencing data.}
\description{
A post hoc cell type classification tool to fine-tune cell type
annotations generated by any cell type classification procedure with
semi-supervised learning algorithm AdaSampling technique.

The current version of scReClassify supports Support Vector Machine and
Random Forest as a base classifier.
}
\seealso{
Useful links:
\itemize{
\item Vignette available at: https://sydneybiox.github.io/scdney/
}
}
\author{
\strong{Maintainer}:
\itemize{
\item Taiyun Kim (ORCID:0000-0002-5028-836X)
\itemize{
\item Email: taiyun.kim91@gmail.com
}
}

Authors:
\itemize{
\item Pengyi Yang (ORCID: 0000-0003-1098-3138)
}
}
