% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bAccuracy.R
\name{bAccuracy}
\alias{bAccuracy}
\title{bAccuracy}
\usage{
bAccuracy(truth, final)
}
\arguments{
\item{cls.truth}{A vector true class label.}

\item{final}{A vector of final classified label prediction from XXX.}
}
\value{
An accuracy value.
}
\description{
This function calculates the accuracy of the prediction to the true label.
}
\author{
Pengyi Yang
}
