% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\alias{gse87795_subset_sce}
\title{GSE827795 subset data}
\format{
An object of class \code{SingleCellExperiment} with 1000 rows and 367 columns.
}
\usage{
gse87795_subset_sce
}
\description{
A SingleCellExperiment object containing a subset expression
matrix of GSE827795. The data contains log2 transformed FPKM expression.

GSE87795 is a mouse fetal liver development data containing 1000 genes,
367 cells and 6 cell types.

The original GSE87795 data and the study details can be found at this
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE87795}{link}
}
\keyword{datasets}
