% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matPCs.R
\name{matPCs}
\alias{matPCs}
\title{matPCs function}
\usage{
matPCs(dat)
}
\arguments{
\item{data}{Prediction model generated from "multiAdaSampling".}

\item{percentVar}{A dimension reduced matrix.}

\item{seed}{seed for randomisation.}
}
\value{
Dimensionally reduced matrix.
}
\description{
Performs PCA on a given matrix
}
\details{
This function performs PCA to reduce the dimension of the gene expression matrix limited from 10 to 20 PCs.
}
\examples{
data("GSE87795_liver.development.data")

mat.expr = GSE87795_liver.development.data$data

mat.pc = matPCs(mat.expr)

# to capture at least 70\% of overall variability in the dataset,
mat.dim.reduct.70 = matPCs(mat.expr, 0.7)
}
\author{
Pengyi Yang, Taiyun Kim
}
